apps/api/pyproject.toml
[project]
name = "wayforth-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.12",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.46.0",
]
